fix compiler warnings, upd print floating point milliseconds

+++ jni/common.hpp
@@ -14,7 +14,7 @@ void doLog(bool shouldPrint, const std::string fmt, ...);
 void diffclock(char const *title, clock_t clock2);
 void difftime(char const *title, std::chrono::time_point<std::chrono::steady_clock> t2, bool shouldExecute);
 void doLogClock(const char* format, const char* title, double diffms);
-void doLogClock1(const char* format, const char* title, long int diffms);
+void doLogClock1(const char* format, const char* title, double diffms);
 void imshowWrapper(const char* name, cv::Mat mat, bool shouldShow);
 void printStatus();
